Rename crate::internal::generics to crate::internal::utils
