--- serde_arrow/src/utils.rs
@@ -1,2 +1,0 @@
-//! Helpers that may be useful when using `serde_arrow`
-pub use crate::internal::generic::{Item, Items};
